Move StorageType trait to common library. Add more logging
