+++ archive/archivist/docker-compose.yml
@@ -11,10 +11,6 @@ s3:
   image: scality/s3server:mem-latest
   ports:
     - "33333:8000"
-dynamodb:
-  image: peopleperhour/dynamodb
-  ports:
-    - "45678:8000"
 localstack:
   image: localstack/localstack
   ports:
