+++ archive/archivist/src/main/scala/uk/ac/wellcome/platform/archive/archivist/models/ArchivistConfig.scala
@@ -40,8 +40,3 @@ case class BagUploaderConfig(
   bagItConfig: BagItConfig = BagItConfig(),
   parallelism: Int
 )
-
-sealed trait StorageType
-case object DigitisedStorageType {
-  override def toString = "digitised"
-}
