+++ archive/archivist/src/test/scala/uk/ac/wellcome/platform/archive/archivist/bag/ArchiveJobCreatorTest.scala
@@ -2,8 +2,8 @@ package uk.ac.wellcome.platform.archive.archivist.bag
 import org.scalatest.{FunSpec, Inside, Matchers}
 import uk.ac.wellcome.platform.archive.archivist.fixtures.ZipBagItFixture
 import uk.ac.wellcome.platform.archive.archivist.generators.BagUploaderConfigGenerator
-import uk.ac.wellcome.platform.archive.archivist.models.{ArchiveJob, BagItConfig, BagManifestLocation, DigitisedStorageType}
-import uk.ac.wellcome.platform.archive.common.models.{BagLocation, BagPath}
+import uk.ac.wellcome.platform.archive.archivist.models.{ArchiveJob, BagItConfig, BagManifestLocation}
+import uk.ac.wellcome.platform.archive.common.models.{BagLocation, BagPath, DigitisedStorageType}
 import uk.ac.wellcome.storage.fixtures.S3.Bucket
 
 class ArchiveJobCreatorTest extends FunSpec with ZipBagItFixture with BagUploaderConfigGenerator with Matchers with Inside{
