+++ archive/common/src/main/scala/uk/ac/wellcome/platform/archive/common/models/StorageType.scala
@@ -1,0 +1,6 @@
+package uk.ac.wellcome.platform.archive.common.models
+
+sealed trait StorageType
+case object DigitisedStorageType extends StorageType {
+  override def toString = "digitised"
+}
